Move api.ts file to app folder
