--- frontend/src/api.ts
@@ -1,23 +1,0 @@
-import axios from "axios";
-import { getAccessToken } from "./features/auth/utils.ts";
-
-const api = axios.create({
-	baseURL: import.meta.env.VITE_API_URL,
-});
-
-api.interceptors.request.use(
-	(config) => {
-		const accessToken = getAccessToken();
-		if (accessToken) {
-			config.headers.Authorization = `Bearer ${accessToken}`;
-		}
-
-		return config;
-	},
-
-	(error) => {
-		return Promise.reject(error);
-	}
-);
-
-export default api;
